requirements.txt
flask
pillow
transformers
python-dotenv
einops
